Add files via upload
About the Project

This data set is about a year's worth of sales from a made up pizza place, including the date and time of each order and the pizzas, served, with additional details on the type, size, quantity, price, and ingredients.

Recommended Analysis:

How many customers do we have each day? Are there any peak hours? How many pizzas are typically in order? Do we have any bestsellers? Can we identify any seasonality in the sales? Are there any pizzas we should take off the menu, or any promotions we could leverage?

Insights:

The peak hours are from 12:00 PM to 1:00 PM on weekdays and 6:00 PM to 8:00 PM on Fridays and Saturdays. The best seller is the Classic Deluxe pizza by order and the Thai Chicken pizza by revenue. Greek pizza in the size XXL is the most expensive and has the lowest orders. By average price and size, there are no orders for Big Meat Pizza in large and medium sizes. The same applies to the Five Cheese Pizza in medium and small, and Four Cheese Pizza in small sizes.